requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-API==2.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
scikit-surprise==1.1.1
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.22
threadpoolctl==2.1.0
urllib3==1.26.2
Werkzeug==1.0.1
